requirements.txt
fastapi
uvicorn
selenium
openai
pydantic
python-dotenv